+++ crates/installer/src/download/mod.rs
@@ -1,0 +1,7 @@
+//! Download system implementation
+//!
+//! This module contains the core downloading functionality, organized by concern:
+//! - `sources/`: Download source implementations
+//! - Core types and traits are in `crate::downloader::core`
+
+pub mod sources;
