+++ crates/installer/src/lib.rs
@@ -4,46 +4,69 @@
 //! for installer applications. It supports multiple download sources,
 //! file integrity validation, retry logic, and progress tracking.
 //!
-//! # Quick Start
+//! # Quick Start - Simple File Download
+//!
+//! ```rust,no_run
+//! use installer::{DownloadRequest, EnhancedDownloader, DownloadConfigBuilder};
+//!
+//! # async fn simple_example() -> installer::Result<()> {
+//! let downloader = EnhancedDownloader::new(DownloadConfigBuilder::new().build());
+//! let request = DownloadRequest::new_http("https://example.com/file.zip", "./downloads");
+//! let result = downloader.download(request, None).await?;
+//! println!("Downloaded: {:?}", result);
+//! # Ok(())
+//! # }
+//! ```
+//!
+//! # Quick Start - Modlist Download (High-level API)
+//!
+//! ```rust,no_run
+//! use installer::{ModlistDownloadBuilder, DashboardStyle};
+//!
+//! # async fn modlist_example() -> installer::Result<()> {
+//! let result = ModlistDownloadBuilder::new("Baseline/modlist")
+//!     .destination("./downloads")
+//!     .automated_only() // Only download what doesn't require user interaction
+//!     .high_performance() // Use all available CPU cores for validation
+//!     .with_dashboard_progress() // Built-in progress reporting
+//!     .max_concurrent_downloads(8)
+//!     .download()
+//!     .await?;
+//!
+//! println!("Downloaded {} files, {} failed", result.successful_downloads, result.failed_downloads);
+//! # Ok(())
+//! # }
+//! ```
+//!
+//! # Low-level API for Custom Use Cases
 //!
 //! ```rust,no_run
 //! use installer::{
 //!     DownloadConfigBuilder, DownloadRequest, EnhancedDownloader,
-//!     FileValidation, ProgressEvent, ConsoleProgressReporter, IntoProgressCallback
+//!     FileValidation, DashboardProgressReporter
 //! };
-//! use std::sync::Arc;
 //!
-//! # async fn example() -> installer::Result<()> {
-//! // Create a download configuration using the builder
+//! # async fn advanced_example() -> installer::Result<()> {
+//! // High-performance configuration
 //! let config = DownloadConfigBuilder::new()
 //!     .high_performance()
 //!     .timeout(std::time::Duration::from_secs(60))
 //!     .build();
 //!
-//! // Create the downloader
 //! let downloader = EnhancedDownloader::new(config);
 //!
-//! // Set up file validation (optional)
+//! // File with validation
 //! let validation = FileValidation::new()
-//!     .with_xxhash64_base64("ejsZTQOI370=")  // Base64 xxHash64 hash
+//!     .with_xxhash64_base64("ejsZTQOI370=")
 //!     .with_expected_size(1024);
 //!
-//! // Create a download request
-//! let request = DownloadRequest::new_http(
-//!     "https://example.com/file.zip",
-//!     "/path/to/download/directory"
-//! )
-//! .with_filename("file.zip")
-//! .with_mirror_url("https://mirror.example.com/file.zip")
-//! .with_validation(validation);
+//! let request = DownloadRequest::new_http("https://example.com/file.zip", "./downloads")
+//!     .with_filename("file.zip")
+//!     .with_validation(validation);
 //!
-//! // Set up progress reporting (optional)
-//! let progress_reporter = ConsoleProgressReporter::new(true); // verbose output
-//! let progress_callback = Some(progress_reporter.into_callback());
-//!
-//! // Download the file
-//! let result = downloader.download(request, progress_callback).await?;
-//! println!("Download result: {:?}", result);
+//! // Built-in dashboard progress reporter
+//! let progress_reporter = DashboardProgressReporter::new();
+//! let result = downloader.download(request, Some(progress_reporter.into_callback())).await?;
 //! # Ok(())
 //! # }
 //! ```
@@ -61,6 +84,8 @@
 
 pub mod downloader;
 pub mod parse_wabbajack;
+pub mod download;
+pub mod integrations;
 
 // Re-export commonly used types for convenience
 pub use downloader::{
@@ -98,6 +123,18 @@ pub use parse_wabbajack::{
     operation_to_download_request, operations_to_download_requests,
     manifest_to_download_requests, manifest_to_prioritized_download_requests,
     manifest_to_download_requests_with_stats, ConversionStats,
+};
+
+// Re-export high-level convenience APIs (the main improvement!)
+pub use integrations::{
+    // Fluent modlist API
+    ModlistDownloadBuilder, ModlistOptions, ModlistDownloadResult, EnhancedDownloaderExt,
+
+    // Built-in progress reporters
+    DashboardProgressReporter, DashboardStyle,
+
+    // Extension traits for better ergonomics
+    DownloadRequestExt, DownloadRequestIteratorExt, DownloadRequestVecExt, RequestSummaryStats,
 };
 
 pub fn add(left: u64, right: u64) -> u64 {
