+++ crates/installer/src/parse_wabbajack/parser.rs
@@ -173,7 +173,6 @@ impl ModlistParser {
             archive.hash.clone(),
             archive.size,
         )
-        .with_hash_algorithm(&self.default_hash_algorithm)
         .with_priority(index as u32) // Use index as default priority
         .with_metadata(metadata);
